+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## 1.1.0 - TBD
+
+- Removed Movement/Patterns from Hexital/Indicator
 - Corrected correlation_coefficient
 - Altered Indicator name reading, to avoid issue's such as "TR" in "ATR" 
 - Added Counter Indicator